requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
pydantic[dotenv]==2.8.2
torch==2.3.1
torchvision==0.18.1
transformers==4.43.3
tokenizers==0.19.1
gradio==3.35.0        